ShiftEv/application/mypagination.py
from rest_framework.pagination import PageNumberPagination
#
# class MyPageNumberPagination(PageNumberPagination):
#     page_size=1
#     # page_query_param='p'
#     # page_size_query_param = 'records'
#     # max_page_size=5


class PaginationHandlerMixin(object):
    @property
    def paginator(self):
        if not hasattr(self, '_paginator'):
            if self.pagination_class is None:
                self._paginator = None
            else:
                self._paginator = self.pagination_class()
        else:
            pass
        return self._paginator
    def paginate_queryset(self, queryset):
        if self.paginator is None:
            return None
        return self.paginator.paginate_queryset(queryset,
                                                self.request, view=self)

    def get_paginated_response(self, data):
        assert self.paginator is not None
        return self.paginator.get_paginated_response(data)