ShiftEv/requirements.txt
aiobotocore==1.4.1
aiohttp==3.7.4.post0
aioitertools==0.8.0
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
boto3==1.18.44
botocore==1.21.65
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.6
django-colorfield==0.4.3
django-cors-headers==3.9.0
django-filter==1.1.0
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
fsspec==2021.8.1
future==0.18.2
html5lib==1.1
idna==3.2
jmespath==0.10.0
jwcrypto==1.0
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.3
Pillow==8.3.2
pkg_resources==0.0.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
pyxlsb==1.0.8
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
s3fs==2021.8.1
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
yarl==1.6.3
